# backend/chronos/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Chronos API
  description: API для планирования встреч с возможностью выбора удобных дат
  version: 1.0.0
  contact:
    name: Chronos Team
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер разработки
...
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      summary: Войти в систему через Яндекс
      description: |
        Авторизация пользователя через токен Yandex OAuth.
        Возвращает JWT токен для доступа к защищенным endpoints.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Невалидный токен Яндекса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /auth/me:
    get:
      tags:
        - auth
      summary: Получить информацию о текущем пользователе
      description: |
        Возвращает информацию о текущем авторизованном пользователе.
        Требует JWT токен в заголовке Authorization.
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Не авторизован
        '403':
          description: Токен недействителен
...
  /meetings/{meetingId}/participants/{participantId}/availability:
    put:
      tags:
        - availability
      summary: Обновить доступность участника
      description: |
        Обновляет доступность участника для конкретной даты.
        
        **Правила авторизации:**
        - Авторизованный пользователь может обновлять только свою доступность
        - Для гостей (неавторизованных) - проверка по participantId из URL
      operationId: updateParticipantAvailability
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: participantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAvailabilityRequest'
      responses:
        '200':
          description: Доступность успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Пользователь пытается изменить чужую доступность
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Встреча или участник не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /meetings/{meetingId}/availability:
    get:
      tags:
        - availability
      summary: Получить всю доступность для встречи
      description: Возвращает всю доступность всех участников встречи
      operationId: getAvailabilityByMeetingId
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Доступность участников
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailabilityResponse'
        '404':
          description: Встреча не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /meetings/{meetingId}/common-dates:
    get:
      tags:
        - availability
      summary: Получить общие доступные даты
      description: Возвращает даты, которые подходят всем участникам встречи
      operationId: getCommonAvailableDates
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Общие доступные даты
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: date
        '404':
          description: Встреча не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен полученный через /auth/login.
        Формат: Bearer <JWT_TOKEN>
  schemas:
...
    LoginRequest:
      type: object
      required:
        - yandexToken
      properties:
        yandexToken:
          type: string
          minLength: 1
          description: Токен авторизации от Yandex OAuth
          example: "test-token"
    LoginResponse:
      type: object
      required:
        - token
        - user
      properties:
        token:
          type: string
          description: JWT токен для авторизации в API
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/UserInfo'
    UserInfo:
      type: object
      required:
        - id
        - email
        - name
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор пользователя
          example: "b08774b1-097b-4d64-8448-1a05b4d49ffb"
        email:
          type: string
          format: email
          description: Email пользователя от Яндекса
          example: "test@yandex.ru"
        name:
          type: string
          description: Имя пользователя
          example: "Тестовый Пользователь"
        avatarUrl:
          type: string
          format: uri
          nullable: true
          description: URL аватара пользователя
          example: "https://avatars.yandex.net/get-yapic/12345/islands-200"

    # ============================================
    # ВСТРЕЧИ - REQUEST/RESPONSE DTOs
    # ============================================
...
    UpdateAvailabilityRequest:
      type: object
      required:
        - date
      properties:
        date:
          type: string
          format: date
          description: Дата доступности
          example: "2024-03-15"
        timeFrom:
          type: string
          format: time
          description: Время начала доступности
          example: "09:00:00"
        timeTo:
          type: string
          format: time
          description: Время окончания доступности
          example: "17:00:00"
...
    AvailabilityResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор доступности
        participantId:
          type: string
          format: uuid
          description: Идентификатор участника
        meetingId:
          type: string
          format: uuid
          description: Идентификатор встречи
        date:
          type: string
          format: date
          description: Дата доступности
        timeFrom:
          type: string
          format: time
          nullable: true
          description: Время начала доступности
        timeTo:
          type: string
          format: time
          nullable: true
          description: Время окончания доступности
        createdAt:
          type: string
          format: date-time
          description: Дата создания записи
...
    ApiErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Сообщение об ошибке
        details:
          type: string
          nullable: true
          description: Дополнительные детали ошибки
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
        path:
          type: string
          nullable: true
          description: Путь запроса, где произошла ошибка